services/gateway/requirements.txt
# Gateway Service Dependencies
# Only API and gRPC client libraries - no ML dependencies

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0

# gRPC Client
grpcio>=1.59.0
protobuf>=4.24.0

# Logging
loguru>=0.7.2

# FastAPI and WebSocket
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Monitoring
prometheus-client>=0.19.0
